Adds project toml and fixes up one dimensional adv-diff tests
