+++ verification/convergence_tests/ConvergenceTests/OneDimensionalCosineAdvectionDiffusion.jl
@@ -2,7 +2,7 @@ module OneDimensionalCosineAdvectionDiffusion
 
 using Printf, Statistics
 
-using Oceananigans, Oceananigans.OutputWriters
+using Oceananigans, Oceananigans.OutputWriters, Oceananigans.Grids
 
 include("analysis.jl")
 
@@ -34,7 +34,8 @@ function run_test(; Nx, Δt, stop_iteration, U = 1, κ = 1e-4,
     println("Running 1D in x cosine advection diffusion test for v and c with Nx = $Nx and Δt = $Δt...")
     run!(simulation)
 
-    c_analytical = c.(grid.xC, 0, 0, model.clock.time, U, κ)
+    x = xnodes(model.tracers.c)
+    c_analytical = c.(x, 0, 0, model.clock.time, U, κ)
 
     # Calculate errors
     cx_simulation = model.tracers.c
